requirements.txt
psycopg2-binary
requests
pyyaml
streamlit
pandas